feat!: use int64 for ID fields (#282)
The `int` datatype is not guaranteed to offer 64 bits of precision. On
32-bit platforms it only has 32 bits of precision. This is incompatible
with the Hetzner Cloud identifiers having up to 52-bits after September
1st 2023.

See #263 for more details about this change.

+++ hcloud/schema/datacenter.go
@@ -2,13 +2,13 @@ package schema
 
 // Datacenter defines the schema of a datacenter.
 type Datacenter struct {
-	ID          int      `json:"id"`
+	ID          int64    `json:"id"`
 	Name        string   `json:"name"`
 	Description string   `json:"description"`
 	Location    Location `json:"location"`
 	ServerTypes struct {
-		Supported []int `json:"supported"`
-		Available []int `json:"available"`
+		Supported []int64 `json:"supported"`
+		Available []int64 `json:"available"`
 	} `json:"server_types"`
 }
